:construction: A homepage mixin for sharing logic
